Refactored in order in the spirit of issue #8
- Parts not essential to data-structure have been moved to Extensions
- Dataframe.lua has changed name to main.lua to avoid confusion regarding loading
- rockspeck now contains all the modules
- The init.lua traverses the extension directory and loads all it finds
- All tests now load the init.lua
- Utils move to a separate file
