+++ tests/categorical_tests.lua
@@ -1,5 +1,6 @@
-require('lfs')
-if (string.match(lfs.currentdir(), "/test")) then
+-- Make sure that directory structure is always the same
+require 'lfs'
+if (string.match(lfs.currentdir(), "/tests$")) then
   lfs.chdir("..")
 end
 paths.dofile('init.lua')
